fix: windows installer and blocking reliability
Summary:
The first attempt at a windows installer had majors issues that resulted
in windows failing to resolve any dns responses and locked the user from
fixing the issues easily. This changed the windows script to use other
open source software that provides a far more stable dns resolution in
windows as a service. The new script will automatically setup the
necessary tools in windows for you. This brings in adguards dnsproxy
(apache license 2.0) and NSSM (public domain) windows binaries so that
they can be used to setup CalypsDoH on windows. The other major fix is
how annoyances and alarmables are matched. Previously, the grep
implementation would do partial matches which resulted in false
positives and over blocking. The change fixes the grep calls to match
exact domains against the blocklists.

Test Plan:
used dnsleaktest.org for its web based doh testing interface to ensure
that both sites that should be blocked are and that a known false
positive is no longer blocked.